Use the same code structure for valves
A class to wrap valve structures has been created and it is instantiated
by means of a factory

Also, a new Map to handle aux nodes, those that encloses a valve or a pump has been defined. Those nodes are added to the Network in the prepare() method. That is a "just in time" strategy.